requirements.txt
streamlit
google-cloud-vision
Pillow
